+++ src/components/button/button.module.css.d.ts
@@ -1,13 +1,15 @@
 /*
- * Copyright (c) 2011-2020 Genestack Limited
+ * Copyright (c) 2011-2021 Genestack Limited
  * All Rights Reserved
  * THIS IS UNPUBLISHED PROPRIETARY SOURCE CODE OF GENESTACK LIMITED
  * The copyright notice above does not evidence any
  * actual or intended publication of such source code.
  */
 export const root: string;
-export const icon: string;
-export const singleIcon: string;
+export const iconStart: string;
+export const iconEnd: string;
+export const iconStartSingle: string;
+export const iconEndSingle: string;
 export const ghost: string;
 export const accent: string;
 export const nowrap: string;